Planteo colisiones NPCs

+++ common/include/DTO/Events/NPCMovedEventDTO.h
@@ -12,7 +12,7 @@ private:
 
 public:
   EventOpcode getCode() const { return opCode; }
-  uint16_t npcId;
+  uint32_t npcId;
   int16_t x;
   int16_t y;
   Direction direction;
